DONE (Minus Noise)
Haven't figured out the noise portion yet but otherwise all parts are complete

+++ Project5/Core/Inc/signal.h
@@ -23,7 +23,7 @@ typedef enum sigtype {
 typedef struct signal {
 	_Bool channel;
 	sigType_t type;
-	double freq;
+	int freq;
 	double minV;
 	double maxV;
 	int noise;
